refactor(GH-9): :recycle: split registry rules files
Original ask: review the repository for the roughly 300-line source-file readability rule, make necessary behavior-preserving reorganizations, and add the work to the existing pull request.

Implementation: split the registry rules fetch, selector, render, apply, merge, markdown, and state helpers into focused files while preserving the existing registry behavior.

Verification: ran go test -count=1 ./pkg/cli ./internal/config; go test -count=1 ./...; go vet ./...; go run ./cmd/kit check --project; git diff --check; go run ./cmd/kit init --refresh --dry-run --diff --file=docs/references/rules/github-pr-delivery.md; go run ./cmd/kit rules list; scanned touched files for common secret patterns.

References: refs #9.
